Preliminary PR feedback
- Remove added argument validation
- Split up tests
- Start to investigate fixing race conditions in the tests (this isn't
done yet)
